OO-Ified 2.0
Moved several files to different packages, cleaned up JIDE.java, and
added an executive driver application for JIDE: JIDE_Exec.
