+++ jIDE/bin/.gitignore
@@ -1,1 +1,2 @@
 /core/
+/keylisteners/
