+++ jIDE/src/core/main/JIDE_Exec.java
@@ -1,0 +1,9 @@
+package core.main;
+
+import core.JIDE;
+
+public class JIDE_Exec {
+	public static void main(String args[]) {
+		new JIDE();
+	}
+}
