+++ jIDE/src/core/system/FileManager.java
@@ -1,4 +1,4 @@
-package core;
+package core.system;
 
 import java.awt.Toolkit;
 import java.io.FileReader;
@@ -7,6 +7,8 @@ import java.io.IOException;
 
 import javax.swing.JFileChooser;
 import javax.swing.JOptionPane;
+
+import core.JIDE;
 
 public class FileManager {
 	static JIDE jideInstance;
